refactor: deduplicate shared utilities across modal/view components (#431)
Phase 1 of view hierarchy flattening — pure deduplication with no behavior changes.

New shared modules:
- lib/formatting.ts: formatBytes() extracted from 4 views
- lib/installHelpers.ts: disk space checker (with auto-cleanup via
  onUnmounted), path guardrails, telemetry helpers extracted from
  NewInstallModal/QuickInstallModal
- composables/useModalOverlay.ts: escape-to-close + click-outside-to-close
  extracted from all 7 modal views, both gated by shouldClose() callback

Updated consumers:
- DetailModal: uses formatBytes import + useModalOverlay
- NewInstallModal: uses formatBytes, installHelpers, useModalOverlay
- QuickInstallModal: uses shared variants lib (removes local dupes),
  formatBytes, installHelpers, useModalOverlay
- ConsoleModal, TrackModal, LoadSnapshotModal, ProgressModal: useModalOverlay

Amp-Thread-ID: https://ampcode.com/threads/T-019db7b6-3400-7293-a487-351dcca1977b
