+++ src/renderer/src/lib/formatting.ts
@@ -1,0 +1,4 @@
+export function formatBytes(bytes: number): string {
+  if (bytes >= 1073741824) return `${(bytes / 1073741824).toFixed(1)} GB`
+  return `${(bytes / 1048576).toFixed(0)} MB`
+}
